requirements.txt
boto==2.42.0
Collectfast==0.2.3
Django==1.10.1
django-activity-stream==0.6.3
django-allauth==0.27.0
django-braces==1.9.0
django-crispy-forms==1.6.0
django-guardian==1.4.6
django-invitations==1.7
django-jsonfield==1.0.1
django-model-utils==2.6
django-notifications-hq==1.2
django-preserialize==1.1.0
django-storages==1.5.1
djangorestframework==3.4.7
docopt==0.4.0
ecdsa==0.13
Fabric==1.12.0
fabric-virtualenv==0.3.0
gunicorn==19.6.0
jsonfield==1.0.3
mailchimp==2.0.9
oauthlib==2.0.0
paramiko==1.17.2
Pillow==3.4.1
psycopg2==2.6.2
pycrypto==2.6.1
python-dateutil==2.5.3
python-openid==2.2.5
pytz==2016.6.1
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
stripe==1.41.0
